Todos CRUD feito, mas somente Jogador no fluxo de service

+++ projeto-campeonato-volei/src/main/java/com/example/projeto/campeonato/volei/repository/JogadorRepository.java
@@ -3,5 +3,5 @@ package com.example.projeto.campeonato.volei.repository;
 import com.example.projeto.campeonato.volei.domain.Jogador;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JogadorRepository extends JpaRepository<Jogador, Long> {
+public interface JogadorRepository extends JpaRepository<Jogador, Integer> {
 }
